refactor: move ui components to own folder
