+++ src/components/ui/CollapseUi/CollapseUi.vue
@@ -17,7 +17,7 @@
 
 <script>
 
-import IconUi from '@/components/ui/IconUi'
+import IconUi from '@/components/ui/IconUi/IconUi'
 
 export default {
   name: 'CollapseUi',
