+++ src/components/ui/MenuUi/MenuUi.vue
@@ -26,7 +26,7 @@
 </template>
 
 <script>
-import IconUi from '@/components/ui/IconUi'
+import IconUi from '@/components/ui/IconUi/IconUi'
 
 export default {
   name: 'MenuUi',
