+++ src/components/ui/index.js
@@ -1,21 +1,21 @@
-import BreadcrumbUi from '@/components/ui/BreadcrumbUi'
-import ButtonUi from '@/components/ui/ButtonUi'
-import ConfirmUi from '@/components/ui/ConfirmUi'
-import FormField from '@/components/ui/form/FormField'
-import IconUi from '@/components/ui/IconUi'
-import MenuUi from '@/components/ui/MenuUi'
-import MessageUi from '@/components/ui/MessageUi'
-import CollapseUi from '@/components/ui/CollapseUi'
-import TabsUi from '@/components/ui/TabsUi'
+import BreadcrumbUi from '@/components/ui/BreadcrumbUi/BreadcrumbUi'
+import ButtonUi from '@/components/ui/ButtonUi/ButtonUi'
+import CollapseUi from '@/components/ui/CollapseUi/CollapseUi'
+import ConfirmUi from '@/components/ui/ConfirmUi/ConfirmUi'
+import FormField from '@/components/ui/form/FormField/FormField'
+import IconUi from '@/components/ui/IconUi/IconUi'
+import MenuUi from '@/components/ui/MenuUi/MenuUi'
+import MessageUi from '@/components/ui/MessageUi/MessageUi'
+import TabsUi from '@/components/ui/TabsUi/TabsUi'
 
 export {
   BreadcrumbUi,
+  CollapseUi,
   ButtonUi,
   ConfirmUi,
   FormField,
   IconUi,
   MenuUi,
   MessageUi,
-  CollapseUi,
   TabsUi
 }
